Font Loading
Create a custom hook called useCachedResources to load
data and resources at called.
Add fonts files to the Assest folder and loaded them
with the useCachedResources at the app open.
Create a custom component for texts with the
loaded fonts.
Refactor the WorkoutItem to use the custom component
with the loaded fonts and passing the styles.